requirements.txt
flask
gunicorn
torch
pillow
opencv-python-headless
pandas
requests
